VNS, framework improvements

+++ src/main/java/es/urjc/etsii/grafo/solver/destructor/Shake.java
@@ -5,5 +5,5 @@ import es.urjc.etsii.grafo.io.Instance;
 import es.urjc.etsii.grafo.solution.Solution;
 
 public interface Shake<S extends Solution<I>, I extends Instance> {
-    void iteration(S s, int k);
+    void iteration(S s, double k);
 }
